chatbot/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
fastmcp==0.3.0
openai==1.12.0
mysql-connector-python==8.2.0
langchain-community==0.0.29
langchain-text-splitters==0.0.1
PyMuPDF==1.23.14
chromadb==0.4.22
pydantic==2.5.3